feat: added MoneyBalanceCard tooltip events; cleaned up moneyEvents file
