+++ app/components/UI/Money/components/MoneyBalanceInfoSheet/MoneyBalanceInfoSheet.tsx
@@ -13,12 +13,21 @@ import { useStyles } from '../../../../../component-library/hooks';
 import styleSheet from './MoneyBalanceInfoSheet.styles';
 import { MoneyBalanceInfoSheetTestIds } from './MoneyBalanceInfoSheet.testIds';
 import { useElevatedSurface } from '../../../../../util/theme/themeUtils';
+import useMountEffect from '../../hooks/useMountEffect';
+import { useMoneyAnalytics } from '../../hooks/useMoneyAnalytics';
+import { BOTTOM_SHEET_NAMES } from '../../constants/moneyEvents';
 
 const MoneyBalanceInfoSheet = () => {
   const sheetRef = useRef<BottomSheetRef>(null);
   const navigation = useNavigation();
   const { styles } = useStyles(styleSheet, {});
   const surfaceClass = useElevatedSurface();
+
+  const { trackBottomSheetViewed } = useMoneyAnalytics({
+    bottom_sheet_name: BOTTOM_SHEET_NAMES.MONEY_BALANCE_INFO_SHEET,
+  });
+
+  useMountEffect(trackBottomSheetViewed);
 
   const handleGoBack = useCallback(() => {
     navigation.goBack();
