+++ app/components/UI/Money/constants/moneyEvents.ts
@@ -1,6 +1,3 @@
-/* eslint-disable @typescript-eslint/consistent-type-definitions */
-
-import { TransactionMeta } from '@metamask/transaction-controller';
 import { METAMASK_SUPPORT_URL } from '../../../../constants/urls';
 import AppConstants from '../../../../core/AppConstants';
 
@@ -31,6 +28,13 @@ export enum BOTTOM_SHEET_NAMES {
   MONEY_EARNINGS_INFO_SHEET = 'money_earnings_info_sheet',
   MONEY_EARN_CRYPTO_INFO_SHEET = 'money_earn_crypto_info_sheet',
   MONEY_MORE_SHEET = 'money_more_sheet',
+  MONEY_BALANCE_INFO_SHEET = 'money_balance_info_sheet',
+}
+
+export enum REDIRECT_TARGETS_TYPES {
+  SCREEN = 'screen',
+  BOTTOM_SHEET = 'bottom_sheet',
+  EXTERNAL_BROWSER = 'external_browser',
 }
 
 // Not using enum because we want to use existing URL constants.
@@ -41,122 +45,65 @@ export const MONEY_URLS = {
 
 export type MONEY_URLS = (typeof MONEY_URLS)[keyof typeof MONEY_URLS];
 
+export enum MONEY_SURFACE_TYPES {
+  SCREEN = 'screen',
+  BOTTOM_SHEET = 'bottom_sheet',
+  COMPONENT = 'component',
+}
+
 export enum COMPONENT_NAMES {
+  // — Section Headers —
+  MONEY_POTENTIAL_EARNINGS_SECTION_HEADER = 'money_potential_earnings_section_header',
+  MONEY_ACTIVITY_SECTION_HEADER = 'money_activity_section_header',
+  MONEY_HOW_IT_WORKS_SECTION_HEADER = 'money_how_it_works_section_header',
+  MONEY_CARD_SECTION_HEADER = 'money_card_section_header',
+
+  // — Onboarding —
+  RIVE_ONBOARDING_STEPPER = 'rive_onboarding_stepper',
+  /** The Stepper Card component on Money Home screen (add funds, get/link card). */
+  MONEY_ONBOARDING_CARD = 'money_onboarding_card',
+
+  // — Earnings —
+  MONEY_ESTIMATED_EARNINGS_SECTION = 'money_estimated_earnings_section',
+  MONEY_POTENTIAL_EARNINGS_SECTION = 'money_potential_earnings_section',
+  MONEY_POTENTIAL_EARNINGS_SECTION_TOKEN_ROW = 'money_potential_earnings_section_token_row',
+  MONEY_POTENTIAL_EARNINGS_TOKEN_ROW = 'money_potential_earnings_token_row',
+
+  // — Activity —
+  MONEY_ACTIVITY_SECTION = 'money_activity_section',
+  MONEY_ACTIVITY_LIST_ITEM = 'money_activity_list_item',
+
+  // — Condensed Info Cards —
+  MONEY_CONDENSED_INFO_CARDS_HOW_IT_WORKS = 'money_condensed_info_cards_how_it_works',
+  MONEY_CONDENSED_INFO_CARDS_MUSD = 'money_condensed_info_cards_musd',
+  MONEY_CONDENSED_INFO_CARDS_WHAT_YOU_GET = 'money_condensed_info_cards_what_you_get',
+
+  // — Add Money Sheet —
+  MONEY_ADD_MONEY_SHEET_CONVERT_CRYPTO = 'money_add_money_sheet_convert_crypto',
+  MONEY_ADD_MONEY_SHEET_DEPOSIT_FUNDS = 'money_add_money_sheet_deposit_funds',
+  MONEY_ADD_MONEY_SHEET_MOVE_MUSD = 'money_add_money_sheet_move_musd',
+
+  // — Transfer Money Sheet —
+  MONEY_TRANSFER_MONEY_SHEET_BETWEEN_ACCOUNTS = 'money_transfer_money_sheet_between_accounts',
+  MONEY_TRANSFER_MONEY_SHEET_PERPS_ACCOUNT = 'money_transfer_money_sheet_perps_account',
+  MONEY_TRANSFER_MONEY_SHEET_PREDICTIONS_ACCOUNT = 'money_transfer_money_sheet_predictions_account',
+
+  // — More Sheet —
+  MONEY_MORE_SHEET_HOW_IT_WORKS = 'money_more_sheet_how_it_works',
+  MONEY_MORE_SHEET_WHAT_YOU_GET = 'money_more_sheet_what_you_get',
+  MONEY_MORE_SHEET_CONTACT_SUPPORT = 'money_more_sheet_contact_support',
+
+  // — Miscellaneous —
+  MONEY_WHAT_YOU_GET_SECTION = 'money_what_you_get_section',
+  MONEY_MUSD_TOKEN_SECTION = 'money_musd_token_row_section',
   MONEY_BALANCE_CARD = 'money_balance_card',
   MONEY_BALANCE_SUMMARY = 'money_balance_summary',
   HOME_TAB = 'home_tab',
   MONEY_ACTION_BUTTON_ROW = 'money_action_button_row',
-  RIVE_ONBOARDING_STEPPER = 'rive_onboarding_stepper',
-  /** The Stepper Card component on Money Home screen (add funds, get/link card). */
-  MONEY_ONBOARDING_CARD = 'money_onboarding_card',
-  MONEY_ESTIMATED_EARNINGS_SECTION = 'money_estimated_earnings_section',
-
-  MONEY_POTENTIAL_EARNINGS_SECTION_HEADER = 'money_potential_earnings_section_header',
-  MONEY_ACTIVITY_SECTION_HEADER = 'money_activity_section_header',
-  MONEY_ACTIVITY_LIST_ITEM = 'money_activity_list_item',
-  MONEY_HOW_IT_WORKS_SECTION_HEADER = 'money_how_it_works_section_header',
-  MONEY_CARD_SECTION_HEADER = 'money_card_section_header',
-
-  MONEY_MUSD_TOKEN_SECTION = 'money_musd_token_row_section',
   MONEY_FOOTER = 'money_footer',
-  MONEY_CONDENSED_INFO_CARDS_HOW_IT_WORKS = 'money_condensed_info_cards_how_it_works',
-  MONEY_CONDENSED_INFO_CARDS_MUSD = 'money_condensed_info_cards_musd',
-  MONEY_CONDENSED_INFO_CARDS_WHAT_YOU_GET = 'money_condensed_info_cards_what_you_get',
-  MONEY_WHAT_YOU_GET_SECTION = 'money_what_you_get_section',
-  MONEY_MORE = 'money_more',
-  MONEY_MORE_SHEET_HOW_IT_WORKS = 'money_more_sheet_how_it_works',
-  MONEY_MORE_SHEET_WHAT_YOU_GET = 'money_more_sheet_what_you_get',
-  MONEY_MORE_SHEET_CONTACT_SUPPORT = 'money_more_sheet_contact_support',
-  MONEY_POTENTIAL_EARNINGS_SECTION = 'money_potential_earnings_section',
-  MONEY_POTENTIAL_EARNINGS_SECTION_TOKEN_ROW = 'money_potential_earnings_section_token_row',
-  MONEY_POTENTIAL_EARNINGS_TOKEN_ROW = 'money_potential_earnings_token_row',
-  MONEY_ADD_MONEY_SHEET_CONVERT_CRYPTO = 'money_add_money_sheet_convert_crypto',
-  MONEY_ADD_MONEY_SHEET_DEPOSIT_FUNDS = 'money_add_money_sheet_deposit_funds',
-  MONEY_ADD_MONEY_SHEET_MOVE_MUSD = 'money_add_money_sheet_move_musd',
-  MONEY_TRANSFER_MONEY_SHEET_BETWEEN_ACCOUNTS = 'money_transfer_money_sheet_between_accounts',
-  MONEY_TRANSFER_MONEY_SHEET_PERPS_ACCOUNT = 'money_transfer_money_sheet_perps_account',
-  MONEY_TRANSFER_MONEY_SHEET_PREDICTIONS_ACCOUNT = 'money_transfer_money_sheet_predictions_account',
-  MONEY_ACTIVITY_SECTION = 'money_activity_section',
   MONEY_CONVERT_CRYPTO_BUTTON = 'money_convert_crypto_button',
+  MONEY_MORE = 'money_more',
 }
-
-export enum REDIRECT_TARGETS_TYPES {
-  SCREEN = 'screen',
-  BOTTOM_SHEET = 'bottom_sheet',
-  EXTERNAL_BROWSER = 'external_browser',
-}
-
-const SCREEN_TARGETS = new Set<string>(Object.values(SCREEN_NAMES));
-const BOTTOM_SHEET_TARGETS = new Set<string>(Object.values(BOTTOM_SHEET_NAMES));
-const URL_TARGETS = new Set<string>(Object.values(MONEY_URLS));
-
-/**
- * Resolves a redirect target's type from the target itself. A target's type is
- * a fact about the target, not something callers should restate, so this is the
- * single source of truth. Fails loud if a target has no category.
- *
- * Resolution is precedence-ordered, so a value duplicated across SCREEN_NAMES,
- * BOTTOM_SHEET_NAMES, or MONEY_URLS would be silently misclassified. Keep these
- * groups disjoint when adding targets (enforced by moneyEvents.test.ts).
- */
-export const resolveRedirectTargetType = (
-  target: SCREEN_NAMES | BOTTOM_SHEET_NAMES | MONEY_URLS,
-): REDIRECT_TARGETS_TYPES => {
-  if (SCREEN_TARGETS.has(target)) return REDIRECT_TARGETS_TYPES.SCREEN;
-  if (BOTTOM_SHEET_TARGETS.has(target)) {
-    return REDIRECT_TARGETS_TYPES.BOTTOM_SHEET;
-  }
-  if (URL_TARGETS.has(target)) return REDIRECT_TARGETS_TYPES.EXTERNAL_BROWSER;
-  throw new Error(
-    `[moneyAnalytics] No redirect_target_type for target: ${target}`,
-  );
-};
-
-// TODO: Breakout types
-/**
- * Properties for tracking location-based events.
- * screen_name: The name of the screen the event occurred on.
- * component_name: The name of the component the event occurred on.
- * bottom_sheet_name: The name of the bottom sheet the event occurred on.
- */
-// TODO: Try to simplify the need for bottom_sheet_name and component_name.
-// Maybe just use surface_name and surface_type?
-export type MoneyLocationEventProperties = {
-  screen_name: SCREEN_NAMES;
-  bottom_sheet_name: BOTTOM_SHEET_NAMES;
-  component_name: COMPONENT_NAMES;
-};
-
-export type MoneyCardEventProperties = {
-  is_card_holder: boolean;
-  is_card_linked_to_money_account: boolean;
-};
-
-type MoneyFundedEventProperties = {
-  is_account_funded: boolean;
-};
-
-export type MoneyRedirectEventProperties = {
-  redirect_target: SCREEN_NAMES | BOTTOM_SHEET_NAMES | MONEY_URLS;
-};
-
-export type MoneySurfaceClickedEventProperties = MoneyRedirectEventProperties &
-  Partial<MoneyLocationEventProperties>;
-
-export type MoneyTokenSurfaceClickedEventProperties =
-  MoneySurfaceClickedEventProperties & MoneyTokenRowEventProperties;
-
-export type MoneyActivitySurfaceClickedEventProperties =
-  MoneySurfaceClickedEventProperties & {
-    transaction: TransactionMeta;
-  };
-
-/**
- * Base properties for all Money events.
- */
-export type MoneyBaseEventProperties = Partial<MoneyLocationEventProperties> &
-  MoneyCardEventProperties &
-  MoneyFundedEventProperties;
 
 /**
  * The intent of the button that was clicked.
@@ -177,47 +124,16 @@ export enum MONEY_BUTTON_TYPES {
   ICON = 'icon',
 }
 
-type MoneyTokenRowEventProperties = {
-  token_symbol: string;
-  token_position_in_list: number;
-  token_chain_id: string;
-  tokens_in_list: number;
-};
+export enum MONEY_TOOLTIP_NAMES {
+  MONEY_BALANCE = 'money_balance',
+  ESTIMATED_EARNINGS = 'estimated_earnings',
+  EARN_ON_YOUR_CRYPTO = 'earn_on_your_crypto',
+  APY = 'apy',
+}
 
-export type MoneyTokenRowButtonClickedEventProperties =
-  MoneyButtonClickedEventProperties & MoneyTokenRowEventProperties;
-
-export type MoneyButtonClickedEventProperties =
-  | ({
-      button_type: MONEY_BUTTON_TYPES.TEXT;
-    } & MoneyTextButtonClickedEventProperties)
-  | ({
-      button_type: MONEY_BUTTON_TYPES.ICON;
-    } & MoneyIconButtonClickedEventProperties);
-
-export type MoneyTextButtonClickedEventProperties = Partial<
-  Pick<MoneyLocationEventProperties, 'component_name'>
-> &
-  Partial<MoneyRedirectEventProperties> & {
-    label_en: string;
-    label_localized: string;
-    button_intent: MONEY_BUTTON_INTENTS;
-    /** 1-based index of the button position in a button row. */
-    button_position?: number;
-    /** Number of buttons in the button row. */
-    button_row_button_count?: number;
-  };
-
-type MoneyIconButtonClickedEventProperties = Partial<
-  Pick<MoneyLocationEventProperties, 'component_name'>
-> &
-  Partial<MoneyRedirectEventProperties> & {
-    button_intent: MONEY_BUTTON_INTENTS;
-    /** 1-based index of the button position in a button row. */
-    button_position?: number;
-    /** Number of buttons in the button row. */
-    button_row_button_count?: number;
-  };
+export enum MONEY_TOOLTIP_TYPES {
+  INFO = 'info',
+}
 
 export enum MONEY_ONBOARDING_STEP_ACTIONS {
   // Generic actions
@@ -232,34 +148,4 @@ export enum MONEY_ONBOARDING_STEP_ACTIONS {
   // Card actions
   GET_CARD = 'get_card',
   LINK_CARD = 'link_card',
-}
-
-export type MoneyOnboardingEventProperties =
-  Partial<MoneyRedirectEventProperties> & {
-    step: number;
-    step_title: string;
-    step_action?: MONEY_ONBOARDING_STEP_ACTIONS;
-    total_steps: number;
-  };
-
-export enum MONEY_TOOLTIP_NAMES {
-  ESTIMATED_EARNINGS = 'estimated_earnings',
-  EARN_ON_YOUR_CRYPTO = 'earn_on_your_crypto',
-  APY = 'apy',
-}
-
-export enum MONEY_TOOLTIP_TYPES {
-  INFO = 'info',
-}
-
-export type MoneyTooltipClickedEventProperties =
-  Partial<MoneyLocationEventProperties> & {
-    tooltip_name: MONEY_TOOLTIP_NAMES;
-    tooltip_type: MONEY_TOOLTIP_TYPES;
-  };
-
-export enum MONEY_SURFACE_TYPES {
-  SCREEN = 'screen',
-  BOTTOM_SHEET = 'bottom_sheet',
-  COMPONENT = 'component',
 }
